refactor(web): extract app bootstrap and fix document IndexedDB save
- Slim App.vue into a shell with composables for platform, sync, and deep links

- Move global styles to assets and bootstrap orchestration to bootstrap.ts

- Serialize post history as plain objects before IndexedDB put
