+++ apps/web/src/assets/less/codemirror-global.less
@@ -1,0 +1,41 @@
+/* CSS-hints */
+.CodeMirror-hints {
+  position: absolute;
+  z-index: 10;
+  overflow-y: auto;
+  margin: 0;
+  padding: 2px;
+  border-radius: 4px;
+  max-height: 20em;
+  min-width: 200px;
+  font-size: 12px;
+  font-family: monospace;
+
+  color: #333333;
+  background-color: #ffffff;
+  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.12), 0 2px 4px 0 rgba(0, 0, 0, 0.08);
+}
+
+.CodeMirror-hint {
+  margin-top: 10px;
+  padding: 4px 6px;
+  border-radius: 2px;
+  white-space: pre;
+  color: #000000;
+  cursor: pointer;
+
+  &:first-of-type {
+    margin-top: 0;
+  }
+  &:hover {
+    background: #f0f0f0;
+  }
+}
+
+.search-match {
+  background-color: #ffeb3b; /* 所有匹配项颜色 */
+}
+
+.current-match {
+  background-color: #ff5722; /* 当前匹配项更鲜艳的颜色 */
+}
